refactor: manager orm store
